requirements.txt
tornado==6.5.2
configparser==7.2.0
paho-mqtt==1.6.1
requests==2.32.5
APScheduler==3.11.0
psutil==7.1.1
xmltodict==1.0.2
openpyxl==3.1.5
PyYAML==6.0.3